Fix for ammo capacity blueprints
The blueprints for the ammo capacity increase of point defence,
heat sink launcher, and chaff launcher have been made more rational
in-game.  Instead of having 3 grades with only grade 3 giving an
extra usable shot, there is now only a grade 1 blueprint that results
in a capacity increase.  This blueprint has the cost of the old g3
blueprint.

These changes have also been applied to coriolis-data with PR 78 to
ensure that the GUIDs still match.